backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
opencv-python==4.8.1.78
numpy==1.26.2
dlib==19.24.2
Pillow==10.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 